feat(config): Rename file*.js to config*.js because they really are config (#13035)
