+++ lib/workers/global/config/parse/__fixtures__/config-async-function.js
@@ -1,4 +1,4 @@
-// This is functionally equivalent to file-function-promise.js but syntactically different
+// This is functionally equivalent to config-function-promise.js but syntactically different
 
 // @ts-ignore
 module.exports = async function () {
