+++ lib/workers/global/config/parse/__fixtures__/config-function-promise.js
@@ -1,4 +1,4 @@
-// This is functionally equivalent to file-async-function.js but syntactically different
+// This is functionally equivalent to config-async-function.js but syntactically different
 
 // @ts-ignore
 module.exports = function () {
