+++ lib/workers/global/config/parse/__fixtures__/config-promise.js
