+++ lib/workers/global/config/parse/__fixtures__/config.js
