+++ lib/workers/global/config/parse/__fixtures__/config2.js
