+++ lib/workers/global/config/parse/file.spec.ts
@@ -2,7 +2,7 @@ import fs from 'fs';
 import { DirectoryResult, dir } from 'tmp-promise';
 import upath from 'upath';
 import { logger } from '../../../../logger';
-import customConfig from './__fixtures__/file';
+import customConfig from './__fixtures__/config';
 import * as file from './file';
 
 describe('workers/global/config/parse/file', () => {
@@ -18,17 +18,17 @@ describe('workers/global/config/parse/file', () => {
 
   describe('.getConfig()', () => {
     it.each([
-      ['custom js config file', 'file.js'],
-      ['custom js config file exporting a Promise', 'file-promise.js'],
-      ['custom js config file exporting a function', 'file-function.js'],
+      ['custom js config file', 'config.js'],
+      ['custom js config file exporting a Promise', 'config-promise.js'],
+      ['custom js config file exporting a function', 'config-function.js'],
       // The next two are different syntactic ways of expressing the same thing
       [
         'custom js config file exporting a function returning a Promise',
-        'file-function-promise.js',
+        'config-function-promise.js',
       ],
       [
         'custom js config file exporting an async function',
-        'file-async-function.js',
+        'config-async-function.js',
       ],
       ['JSON5 config file', 'config.json5'],
       ['YAML config file', 'config.yaml'],
@@ -40,7 +40,7 @@ describe('workers/global/config/parse/file', () => {
     });
 
     it('migrates', async () => {
-      const configFile = upath.resolve(__dirname, './__fixtures__/file2.js');
+      const configFile = upath.resolve(__dirname, './__fixtures__/config2.js');
       const res = await file.getConfig({ RENOVATE_CONFIG_FILE: configFile });
       expect(res).toMatchSnapshot();
       expect(res.rangeStrategy).toBe('bump');
